Fix PHPCS linting violations
- Auto-fixed 14 violations via phpcbf (whitespace, brace placement, spacing)
- Extracted MockCollector into its own file to satisfy PSR-12 one-class-per-file rule
